requirements.txt
asgiref==3.5.1
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
flake8-django==1.1.1
flake8-html==0.4.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
